third_party/WebKit/LayoutTests/editing/pasteboard/copy-cut-paste-supported.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
test(function() {
  assert_true(document.queryCommandSupported('copy'));
  assert_true(document.queryCommandSupported('cut'));
  assert_true(document.queryCommandSupported('paste'));
});
</script>
